requirements.txt
discord
python-dotenv
unalix
